# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ShopFlow ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    get:
      operationId: getCategory
      tags: [categories]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    patch:
      operationId: updateCategory
      tags: [categories]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Brands
  /brands:
    get:
      operationId: listBrands
      tags: [brands]
      summary: List all brands
      responses:
        "200":
          description: List of brands
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Brand"
    post:
      operationId: createBrand
      tags: [brands]
      summary: Create a brand
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BrandInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Brand"
  /brands/{id}:
    patch:
      operationId: updateBrand
      tags: [brands]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BrandUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Brand"
    delete:
      operationId: deleteBrand
      tags: [brands]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Products
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List customers
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Customers list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomersPage"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    patch:
      operationId: updateCustomer
      tags: [customers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /customers/{id}/ledger:
    get:
      operationId: getCustomerLedger
      tags: [customers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer ledger
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LedgerEntry"

  # Suppliers
...
  /employees:
    get:
      operationId: listEmployees
      tags: [employees]
      summary: List employees
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: role
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Employees list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeesPage"
    post:
      operationId: createEmployee
      tags: [employees]
      summary: Create an employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /employees/{id}:
    get:
      operationId: getEmployee
      tags: [employees]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    patch:
      operationId: updateEmployee
      tags: [employees]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    delete:
      operationId: deleteEmployee
      tags: [employees]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Expenses
...
components:
  schemas:
...
    Category:
      type: object
      required: [id, name, createdAt]
      properties:
        id:          { type: integer }
        name:        { type: string }
        description: { type: ["string", "null"] }
        productsCount: { type: integer }
        createdAt:   { type: string }
    CategoryInput:
      type: object
      required: [name]
      properties:
        name:        { type: string, minLength: 1 }
        description: { type: string }
    CategoryUpdate:
      type: object
      properties:
        name:        { type: string, minLength: 1 }
        description: { type: string }
    Brand:
      type: object
      required: [id, name, createdAt]
      properties:
        id:          { type: integer }
        name:        { type: string }
        description: { type: ["string", "null"] }
        productsCount: { type: integer }
        createdAt:   { type: string }
    BrandInput:
      type: object
      required: [name]
      properties:
        name:        { type: string, minLength: 1 }
        description: { type: string }
    BrandUpdate:
      type: object
      properties:
        name:        { type: string, minLength: 1 }
        description: { type: string }
...
    Customer:
      type: object
      required: [id, name, phone, customerCode, createdAt]
      properties:
        id:           { type: integer }
        customerCode: { type: string }
        name:         { type: string }
        phone:        { type: string }
        email:        { type: ["string", "null"] }
        address:      { type: ["string", "null"] }
        gstNumber:    { type: ["string", "null"] }
        creditLimit:  { type: number }
        outstanding:  { type: number }
        type:         { type: string }
        status:       { type: string }
        notes:        { type: ["string", "null"] }
        createdAt:    { type: string }
    CustomersPage:
      type: object
      required: [data, total, page, limit]
      properties:
        data:  { type: array, items: { $ref: "#/components/schemas/Customer" } }
        total: { type: integer }
        page:  { type: integer }
        limit: { type: integer }
    CustomerInput:
      type: object
      required: [name, phone]
      properties:
        name:        { type: string, minLength: 1 }
        phone:       { type: string }
        email:       { type: string }
        address:     { type: string }
        gstNumber:   { type: string }
        creditLimit: { type: number }
        type:        { type: string }
        notes:       { type: string }
    CustomerUpdate:
      type: object
      properties:
        name:        { type: string }
        phone:       { type: string }
        email:       { type: string }
        address:     { type: string }
        gstNumber:   { type: string }
        creditLimit: { type: number }
        type:        { type: string }
        status:      { type: string }
        notes:       { type: string }
    LedgerEntry:
      type: object
      required: [id, date, description, debit, credit, balance, type]
      properties:
        id:          { type: integer }
        date:        { type: string }
        description: { type: string }
        debit:       { type: number }
        credit:      { type: number }
        balance:     { type: number }
        type:        { type: string }
        referenceId: { type: ["integer", "null"] }
...
    Employee:
      type: object
      required: [id, name, phone, role, employeeCode, status, joiningDate, createdAt]
      properties:
        id:           { type: integer }
        employeeCode: { type: string }
        name:         { type: string }
        phone:        { type: string }
        email:        { type: ["string", "null"] }
        address:      { type: ["string", "null"] }
        role:         { type: string }
        department:   { type: ["string", "null"] }
        salary:       { type: number }
        status:       { type: string }
        joiningDate:  { type: string }
        notes:        { type: ["string", "null"] }
        createdAt:    { type: string }
    EmployeesPage:
      type: object
      required: [data, total, page, limit]
      properties:
        data:  { type: array, items: { $ref: "#/components/schemas/Employee" } }
        total: { type: integer }
        page:  { type: integer }
        limit: { type: integer }
    EmployeeInput:
      type: object
      required: [name, phone, role, joiningDate, salary]
      properties:
        name:        { type: string, minLength: 1 }
        phone:       { type: string }
        email:       { type: string }
        address:     { type: string }
        role:        { type: string }
        department:  { type: string }
        salary:      { type: number }
        joiningDate: { type: string }
        notes:       { type: string }
    EmployeeUpdate:
      type: object
      properties:
        name:       { type: string }
        phone:      { type: string }
        email:      { type: string }
        address:    { type: string }
        role:       { type: string }
        department: { type: string }
        salary:     { type: number }
        status:     { type: string }
        notes:      { type: string }
...